Updated chebyshev_bestN and added chebfun
Updated the chebyshev_bestN function to compute both the coefficients
and the value of N, by downsampling, that meets our error requirements.

chebfun, a function for creating a chebyshev approximation using
coefficient evaluations is now available which is faster than the
interpolating function for approximations, and is awesome at what it
does.

Also deleted all of the play code at the bottom of the file