Added chebfun, parabolaInterpolate, meroSFunc, and CInfParabolaSmoother
List of updates:
+ chebfun: Return a function that allows one to replace fxn with
chebfun(fxn) and easily evaluate a function to machine precision with a
Chebfun. This function uses the coefficients of a chebyshev
approximation as opposed to the interpolating polynomial and it is very
good at projecting the fxn onto a chebyshev series
+ parabolaInterpolate: return the polynomial interpolate through three
points - this function is hard-coded so it is really fast
+ meroSFunc: A function that takes in x \in [0,1] and returns a weight
from a meromorphic CInfinity function that is infinity differentiable.
It is useful for constructing spline functions using only parabolas
+ CInfParabolaSmoother: This function, with inspiration from Ulrich
Multze, uses a meromorphic CInfinity weight function given by: 1 / (1 +
exp(p * (1/ x + 1 / (x -1)))) with p = 0.810747. This function allows
you to interpolate a set of (x,y) pairs perfectly with minimal work and
quick functional execution
+ Added a new return value for chebyshev_bestN: Instead of returning
just a value of N that satisfies the criteria for the error, this
function now returns the coefficients as well. They are all stored in a
list

- Misc. Code at the bottom